refactor(grfbrowser): Extract utils.go from main.go
Move utility functions to separate file for better organization:
- euckrToUTF8: EUC-KR to UTF-8 string conversion
- sprImageToRGBA: SPR image to RGBA conversion
- getFileIcon: File type icon mapping
- getFileTypeName: Human-readable file type names

Part of ADR-012 Stage 0: GRF Browser refactoring.
